# api/go-client/model_user.go
/*
 * 基于区块链的供应链金融平台
 *
 * Blockchain
 *
 * API version: 1.0.0
 * Contact: chris-ju@qq.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type User struct {
	Username string `json:"username,omitempty"`
	Password string `json:"password,omitempty"`
	Company *Company `json:"company,omitempty"`
}